Move some code and fix import references
